Split the three agent host files that outgrew the size ratchet
The ratchet records pre-existing debt, not exemptions for new code, so these
are splits rather than a raised baseline. Each cut follows a seam that was
already there:

* agent_host_events (630) keeps normalizing; _Segment and the event
  constructors move to agent_host_event_text, next to the wire contract they
  share with the host's Rust `chunk_text`;
* agent_host (696) keeps driving a run; assembling and dispatching one, and
  renewing its credential, move to agent_host_dispatch -- different job,
  different failure modes, sharing only a run id;
* agent_host_dispatch_repository (712) keeps its surface; the rules for
  applying what a host reports move to agent_host_control_updates, following
  agent_host_recovery and agent_host_event_intake.
