+++ lemma-backend/app/modules/agent/infrastructure/agent_host_dispatch_repository.py
@@ -46,7 +46,6 @@ from app.modules.agent.domain.agent_host import (
     AgentHostRunCheckpoint,
     AgentHostRunSpec,
     AgentHostRunState,
-    run_state_progresses,
 )
 from app.modules.agent.infrastructure.agent_host_event_stream import (
     AgentHostEventStream,
@@ -54,17 +53,15 @@ from app.modules.agent.infrastructure.agent_host_event_stream import (
 )
 from app.modules.agent.infrastructure import (
     agent_host_admission,
+    agent_host_control_updates,
     agent_host_event_intake,
     agent_host_recovery,
-    agent_host_session_memory,
 )
 from app.modules.agent.infrastructure.agent_host_repository_common import (
     DEFAULT_COMMAND_TTL_SECONDS,
     DEFAULT_PERMISSION_COMMAND_TTL_SECONDS,
     DEFAULT_RUN_LEASE_SECONDS,
-    AgentHostNotFound,
     AgentHostProtocolViolation,
-    AgentHostRepositoryError,
     utcnow,
 )
 from app.modules.agent.infrastructure.runtime_models import (
@@ -178,12 +175,15 @@ class AgentHostDispatchRepository:
         lease_seconds: int = DEFAULT_RUN_LEASE_SECONDS,
     ) -> PolledCommands:
         timestamp = now or utcnow()
-        acknowledged = await self._acknowledge_commands(
+        acknowledged = await agent_host_control_updates.acknowledge_commands(
+            self.session,
             host_id=host_id,
             command_ids=acknowledged_command_ids,
             now=timestamp,
         )
-        applied = await self._apply_control_updates(
+        applied = await agent_host_control_updates.apply_control_updates(
+            self.session,
+            self.uow,
             host_id=host_id,
             checkpoints=checkpoints,
             rejections=rejections,
@@ -251,272 +251,6 @@ class AgentHostDispatchRepository:
             progressed=bool(acknowledged or applied),
         )
 
-    async def _apply_control_updates(
-        self,
-        *,
-        host_id: UUID,
-        checkpoints: list[AgentHostRunCheckpoint],
-        rejections: list[AgentHostCommandRejection],
-        now: datetime,
-        lease_seconds: int,
-    ) -> int:
-        """Apply the host's reported updates, isolating each one.
-
-        Returns how many of them actually changed something, so a poll carrying
-        nothing but repeated heartbeats can be told apart from one carrying
-        news.
-
-        A failure here is one run's problem and must never become this host's:
-        the poll that carries these updates is also the only way commands reach
-        the host, so raising would stop CANCEL_RUN and RESOLVE_PERMISSION
-        reaching every other run it is executing. The host would then resend the
-        same update on its next poll and wedge itself permanently.
-
-        ValueError is caught alongside the typed failures because a lease row
-        carrying a state this build no longer parses would otherwise raise out
-        of the enum conversion and wedge the host just as effectively.
-        """
-        changed = 0
-        for checkpoint in checkpoints:
-            try:
-                if await agent_host_session_memory.remember_provider_session(
-                    self.uow, checkpoint
-                ):
-                    changed += 1
-                _, advanced = await self._apply_checkpoint(
-                    host_id=host_id,
-                    checkpoint=checkpoint,
-                    now=now,
-                    lease_seconds=lease_seconds,
-                )
-                changed += int(advanced)
-            except (AgentHostRepositoryError, ValueError) as exc:
-                _log_unappliable_update(
-                    kind="checkpoint",
-                    host_id=host_id,
-                    run_id=checkpoint.run_id,
-                    exc=exc,
-                )
-        for rejection in rejections:
-            try:
-                changed += int(
-                    await self.apply_rejection(
-                        host_id=host_id,
-                        rejection=rejection,
-                        now=now,
-                    )
-                )
-            except (AgentHostRepositoryError, ValueError) as exc:
-                _log_unappliable_update(
-                    kind="rejection",
-                    host_id=host_id,
-                    run_id=rejection.run_id,
-                    exc=exc,
-                )
-        return changed
-
-    async def apply_rejection(
-        self,
-        *,
-        host_id: UUID,
-        rejection: AgentHostCommandRejection,
-        now: datetime | None = None,
-    ) -> bool:
-        """Persist one fenced pre-dispatch rejection atomically.
-
-        A receipt can requeue an unaccepted command or terminalize it, but it
-        can never move an accepted lease backwards. Duplicate and stale
-        receipts therefore become harmless no-ops.
-
-        Returns whether this receipt changed anything, so a resent one is not
-        mistaken for news.
-        """
-        timestamp = now or utcnow()
-        command = await self.session.get(
-            AgentHostCommandModel,
-            rejection.command_id,
-            with_for_update=True,
-        )
-        if command is None or command.host_id != host_id:
-            raise AgentHostProtocolViolation(
-                "rejected command does not belong to this host"
-            )
-        if (
-            command.kind != AgentHostCommandKind.START_RUN.value
-            or command.run_id != rejection.run_id
-            or command.lease_epoch != rejection.lease_epoch
-        ):
-            raise AgentHostProtocolViolation(
-                "rejection identity does not match command"
-            )
-        lease = await self.session.get(
-            AgentHostRunLeaseModel,
-            rejection.run_id,
-            with_for_update=True,
-        )
-        if (
-            lease is None
-            or lease.host_id != host_id
-            or lease.lease_epoch != rejection.lease_epoch
-        ):
-            return False
-        if (
-            lease.accepted_at is not None
-            or command.state == AgentHostCommandState.ACKNOWLEDGED.value
-            or AgentHostRunState(lease.state) in TERMINAL_AGENT_HOST_RUN_STATES
-        ):
-            return False
-
-        command.rejection = {
-            "code": rejection.code.value,
-            "retryable": rejection.retryable,
-            "detail": rejection.detail,
-            "rejected_at": timestamp.isoformat(),
-        }
-        if rejection.retryable:
-            command.state = AgentHostCommandState.QUEUED.value
-            command.delivered_at = None
-            lease.state = AgentHostRunState.QUEUED_FOR_HOST.value
-        else:
-            command.state = AgentHostCommandState.ACKNOWLEDGED.value
-            command.acknowledged_at = timestamp
-            lease.state = AgentHostRunState.FAILED.value
-            lease.terminal_at = timestamp
-            lease.error_code = rejection.code.value
-            lease.error_detail = rejection.detail
-        lease.updated_at = timestamp
-        await self.session.flush()
-        return True
-
-    async def _acknowledge_commands(
-        self,
-        *,
-        host_id: UUID,
-        command_ids: list[UUID],
-        now: datetime,
-    ) -> int:
-        """Mark delivered commands done; returns how many actually changed."""
-        if not command_ids:
-            return 0
-        result = await self.session.execute(
-            select(AgentHostCommandModel)
-            .where(
-                AgentHostCommandModel.host_id == host_id,
-                AgentHostCommandModel.id.in_(command_ids),
-            )
-            .with_for_update()
-        )
-        found = {command.id: command for command in result.scalars()}
-        acknowledged = 0
-        for command_id in command_ids:
-            command = found.get(command_id)
-            if command is None:
-                # Nothing to mark: the command was swept by retention, or it
-                # was never ours. Either way the host has already stopped
-                # executing it, and refusing the poll would only make it resend
-                # the same acknowledgement forever.
-                _log_unappliable_update(
-                    kind="acknowledgement",
-                    host_id=host_id,
-                    run_id=None,
-                    exc=AgentHostNotFound(f"command {command_id} is unknown"),
-                )
-                continue
-            if command.state in {
-                AgentHostCommandState.CANCELLED.value,
-                AgentHostCommandState.EXPIRED.value,
-                AgentHostCommandState.ACKNOWLEDGED.value,
-            }:
-                continue
-            command.state = AgentHostCommandState.ACKNOWLEDGED.value
-            command.acknowledged_at = now
-            acknowledged += 1
-        return acknowledged
-
-    async def apply_checkpoint(
-        self,
-        *,
-        host_id: UUID,
-        checkpoint: AgentHostRunCheckpoint,
-        now: datetime | None = None,
-        lease_seconds: int = DEFAULT_RUN_LEASE_SECONDS,
-    ) -> AgentHostRunLeaseModel | None:
-        """Advance a run's state and extend its lease.
-
-        This is the lease heartbeat. Event batches deliberately do not extend
-        it, so an active run is kept alive by its poll cycle rather than by a
-        row write per batch of output.
-
-        Idempotent in exactly the way :meth:`apply_rejection` is, and for the
-        same reason: a checkpoint the host cannot get us to accept is one it
-        resends every poll forever. A checkpoint for a lease we no longer have,
-        for a superseded epoch, for an already-terminal run, or reporting a
-        state behind the one we hold is *information we have already acted on*,
-        not a protocol breach. All four return ``None`` and change nothing.
-
-        The realistic case is not a buggy host: a laptop sleeps, we reconcile
-        the run to RECOVERING and then to the terminal DISPATCH_UNKNOWN, and the
-        laptop wakes up still believing it is RUNNING. Its next checkpoint is
-        both terminal-violating and regressive, and it will resend it until we
-        stop refusing it.
-        """
-        lease, _ = await self._apply_checkpoint(
-            host_id=host_id,
-            checkpoint=checkpoint,
-            now=now,
-            lease_seconds=lease_seconds,
-        )
-        return lease
-
-    async def _apply_checkpoint(
-        self,
-        *,
-        host_id: UUID,
-        checkpoint: AgentHostRunCheckpoint,
-        now: datetime | None = None,
-        lease_seconds: int = DEFAULT_RUN_LEASE_SECONDS,
-    ) -> tuple[AgentHostRunLeaseModel | None, bool]:
-        """:meth:`apply_checkpoint`, also saying whether the state advanced.
-
-        The lease alone cannot answer that: a terminal run re-reporting the
-        terminal state it already holds returns its lease and changes nothing,
-        and a still-RUNNING run re-reporting RUNNING is the heartbeat rather
-        than news. Both extend the lease; neither is a reason to cut a long
-        poll short.
-        """
-        timestamp = now or utcnow()
-        lease = await self.session.get(
-            AgentHostRunLeaseModel,
-            checkpoint.run_id,
-            with_for_update=True,
-        )
-        if lease is None or lease.host_id != host_id:
-            return None, False
-        if lease.lease_epoch != checkpoint.lease_epoch:
-            return None, False
-        current_state = AgentHostRunState(lease.state)
-        reported = checkpoint.state
-        if current_state in TERMINAL_AGENT_HOST_RUN_STATES:
-            return (lease, False) if reported is current_state else (None, False)
-        if not run_state_progresses(current_state, reported):
-            return None, False
-        advanced = reported is not current_state
-        # accepted_at is the single fence between pre-dispatch (safe to retry
-        # or fall back) and accepted (never repeated). Crossing it is news in
-        # its own right, whatever the state did.
-        if lease.accepted_at is None:
-            lease.accepted_at = timestamp
-            advanced = True
-        if reported in TERMINAL_AGENT_HOST_RUN_STATES:
-            lease.terminal_at = timestamp
-        lease.state = reported.value
-        lease.lease_expires_at = timestamp + timedelta(seconds=lease_seconds)
-        lease.updated_at = timestamp
-        await self.session.flush()
-        return lease, advanced
-
-    # ------------------------------------------------------------------ events
-
     async def append_events(
         self,
         *,
@@ -534,6 +268,43 @@ class AgentHostDispatchRepository:
     async def delete_run_events(self, *, run_id: UUID) -> None:
         """Drop a terminalized run's stream."""
         await self._events.delete(run_id=run_id)
+
+    # -------------------------------------------------------- control updates
+    #
+    # The rules live in agent_host_control_updates; these keep one entry point
+    # for callers, as the recovery and intake delegates above do.
+
+    async def apply_checkpoint(
+        self,
+        *,
+        host_id: UUID,
+        checkpoint: AgentHostRunCheckpoint,
+        now: datetime | None = None,
+        lease_seconds: int = DEFAULT_RUN_LEASE_SECONDS,
+    ) -> AgentHostRunLeaseModel | None:
+        """Advance a run's state and extend its lease."""
+        return await agent_host_control_updates.apply_checkpoint(
+            self.session,
+            host_id=host_id,
+            checkpoint=checkpoint,
+            now=now,
+            lease_seconds=lease_seconds,
+        )
+
+    async def apply_rejection(
+        self,
+        *,
+        host_id: UUID,
+        rejection: AgentHostCommandRejection,
+        now: datetime | None = None,
+    ) -> bool:
+        """Persist one fenced pre-dispatch rejection atomically."""
+        return await agent_host_control_updates.apply_rejection(
+            self.session,
+            host_id=host_id,
+            rejection=rejection,
+            now=now,
+        )
 
     # --------------------------------------------------------------- lifecycle
